Better Scene
Coded up a new scene and made a few improvements here are there. Off
axis frustum disabled in this version since I am at home and have no
kinect to test it with and the code needs to be rewritten and better
done now that I have a better understanding.