Bug fixes and features galore
Federal regulations require me to warn you that this next demo... is
looking pretty good.

All joking aside a lot was added in this version to attempt to squash
some bugs that found their way into the code as well as some useful
added features that include the following:
Ability to invert eyes so that right and left eyes are seeing the
correct image. (both image is marked now as to which eye should see it)
Meshes can now be scaled to a unit amount when called which will match
up with where the user thinks it is a real world space.
HUD is now locked in place and will not move with scene
You can now toggle pulling data from kinect or not.
Buttons 8 and 5 can now exaggerate the object as being closer or farther
from screen.
9 and 6 change the distance between two eyes.
Top Bottom and Left Right 3D (Passive 3D) is now supported as well.

Also for future note expect any updates commited to be larger and less
frequent now as my new build machine does not have github. Also there is
another build which will be uploaded likely today since this build I had
intended to upload yesterday which deals with a few more issues that
exist in this version.