Slight Improvements
Mesh importer now has more scaling options all of which are decided
during the original call now instead of on every draw (For use later
when a loadOptions.txt is added)
Changed light on mesh to highlight meshes better.
More pushes and pops to allow drawing the room without the spheres.

After staring at a screen and doing several tests while the 3D is
showing the right size (A 60 cm object looks to the user to be 60 cm
when the screen sizes are correct) it does not match up properly to the
movement of the object. I can get one to be right when modifying during
run time but not both so the frustum and look at may need some more
work.