Version 3
Stable and bug free enough to be worth of a version number.
Major Changes:
Screen size can be set to be different for passive and active display
and will change accordingly at runtime
Lighting fixes now you should not be able to get a different light
amount cast on each eye.
Dynamic scene changing
Attempted to chase down a ghosting bug where right eye would see a
ghosted image of the left eye. I thought I might have caused this but
not clearing the buffers well enough but it would seem after using some
debug tools that this is just a limitation of these displays that is
obvious when drawing a crazy distances in front of screen.
Fullscreen now is toggleable
HUD now works nicer when changing the for loop for each eye (useful for
debugging)