Version 4
A few bug fixes.
Positional cheat works very well now a power point presentation that
covers it a bit will be included in the readme in a few days. Also
better documentation will also be done over the next few days other than
that this project is done for now.